$CONVENTION_NAME
When boilerplate/_lib/common.sh is sourced by a convention `update`
script, it's useful for that script to know the name of its convention
without having to hardcode it. This commit adds logic to common.sh to
discover this and put it in a `$CONVENTION_NAME` variable. The
openshift/golang_osd_cluster_operator convention's `update` script is
modified to use the new variable.

As part of this discovery, `$HERE` (the directory containing the file
doing the sourcing) is also discovered and set. This commit removes it
from places where it is no longer needed (or never was).

Also remove a redundant shebang from a lib that should only ever be
sourced.