Copy golang_osd_cluster_operator to ..._hive_...
At this point in history, all the content of the
openshift/golang_osd_cluster_operator convention is actually applicable
to both cluster-installed and hive-installed operators. So there should
be nothing stopping us from publishing both conventions.

Eventually we're going to have a smarter way to maintain the common bits
in a single place. But for now we're dumbly copying the contents of
golang_osd_cluster_operator into golang_osd_hive_operator.

While these conventions are duplicates of each other, it's important
that they stay in sync. So this commit also adds a test case to enforce
that. The test case is expected to be temporary, to be removed as soon
as the aforementioned "smarter way" is implemented.