One-time "nexus" Makefile include
Previously, any Makefile content in subscribed conventions needed to be
pulled in -- and maintained on each update if the filenames or locations
changed -- by manually curating `include` directives in your base
Makefile.

With this commit, you only need to add

`include boilerplate/generated-includes.mk`

to your base Makefile when you bootstrap (or update to/past this
commit). The contents of that file are regenerated on each update to
automatically `include` any `*.mk` files in your subscribed conventions.