`operator-sdk generate` at the proper version
Enhances ensure.sh to recognize the `operator-sdk` directive. This
discovers the version of operator-sdk used by the consuming repository
(from go.mod) and makes it available via a symlink at
`./.operator-sdk/bin/operator-sdk`. If there's already one installed in
your `$PATH`, and it's the right version, we just symlink to it.
Otherwise we download the appropriate binary.

Adds a script to boilerplate/_lib called `operator-sdk-generate.sh` that
uses the above, and then executes the `operator-sdk generate` commands
appropriate to the version.

Swaps out the guts of `make opgenerate` to invoke the above.