`update` creates a commit
The main `update` driver is modified to create a git commit containing
the changes it brings in.

Because we're committing, it becomes more important that the consuming
checkout is clean before we start, so that it's easy to revert if
something goes wrong. A hard check is added for that. However, we make
an exception for update.cfg so that you don't have to have a whole
commit with just a delta to that file.

In the spirit of the famed and fabled `promote.sh`, the generated
commit's message contains a history of the boilerplate commits since the
last update. But it also handles an initial bootstrap sanely (you don't
get the full history of the boilerplate repo). It also lists the
conventions processed.

A granny switch is provided to avoid the automated commit:
`export BOILERPLATE_UPDATE_NO_COMMIT=1`