Disallow (most) dirty checkouts when updating
Since boilerplate update messes with your local checkout, it can be hard
to recover pre-existing local changes if something goes wrong and you
need to back out. With this commit, update fails quickly if your
checkout is not clean.

However, to facilitate the most common workflows, we allow the following
exceptions:
- We try to detect when you're bootstrapping, and allow only the files
related to that.
- We try to detect when you're only updating the config, i.e. changing
subscriptions, and allow that.