Script/target to generate a commit after update
This adds a `boilerplate-commit` target brought in by the nexus
makefile, intended for use by consuming repositories. The target invokes
a script of the same name that creates a git commit out of your current
unstaged changes if they relate to boilerplate bootstrap, update, and/or
subscriptions. When the changes include an update, the generated commit
message contains a log summary of the boilerplate commits brought in by
the update, as well as a github `compare` URL for the same, in the
spirit of promote.sh.