Review updates from PR #56
- Soften README wording around using `make boilerplate-commit` to make
it sound less mandatory.
- Remove `go.mod`, `go.sum`, and per-convention
`.boilerplate-ignore-files` from the ignore list when ensuring a "clean"
checkout prior to `update`. These were causing a test failure, but the
right thing was to fix the test, which this commit does.
- Add some comments around non-obvious chunks of code.
- Factor out `BOILERPLATE_GIT_REPO` and `BOILERPLATE_GIT_CLONE`
defaulting. (Note that this still isn't used by the main `update`
driver, because that needs to be self-contained for bootstrapping.)
- Use common.sh for freeze-check, as part of the above.