Make go-build only build
Since we're running them in prow, and have made sure the prow
environment matches the appsre environment, there's no need for
validation steps like linting and `go test` to be run as part of the
container build.

This commit makes `make go-build` only do the go build. Since `make
go-build` is the crux of `make docker-build`, this makes the container
build more efficient.

(Note that, prior to #103, `make docker-build` was the only way to run
these validation pieces "locally" in the proper container. Now you can
do it with `boilerplate/_lib/container-make {target}`.)