Move consumer detection to release.sh
Consumer detection (discovering the github org, project, and branch name
of the consuming project) relies on git config being set up a certain
way, which it will be for a human's clone of the repository, but not
always in CI. Inclusion in common.sh was thus causing CI failures. So
move this logic to release.sh, which is only sourced by user-facing
scripts.