Alphabetize .ci-operator.yaml
When `make prow-config` parlays .ci-operator.yaml into (the temporary
section in) the prow configuration, apparently prow's CI cares that the
fields are ordered alphabetically.

Consumers not updated with this commit can work around this by manually
reordering in their release PR.

And this will become n/a anyway once DPTP-1640 is gone.