Remove DPTP-1640 workaround for consumers
It turns out we don't need to do the workaround for DPTP-1640 for every
consumer -- we only need to do it once in the whole environment. That's
now being taken care of in [boilerplate's config](https://github.com/openshift/release/pull/13288),
so this commit removes it from what consumers generate via
`make prow-config`.

DPTP-1640