WIP: Default Dockerfile in openshift/golang-osd-operator
Add a standard Dockerfile to the openshift/golang-osd-operator
convention. Consumers updating to this level should either
- Delete `build/Dockerfile` and define `OPERATOR_DESCRIPTION` in the
local `Makefile` (preferred); or
- Override `OPERATOR_DOCKERFILE` in the local `Makefile`. But only if
you _really_ can't use the standard one.

WIP:
- How are we going to keep the initial `FROM` up to date? We can keep
overriding it from the convention's `update`; or we can hardcode it and
add a check to `03-image-tags` to make sure we don't forget to update it
when publishing a new image.
- Are we going to bother continuing with the `FROM` hack in the
convention's `update` if they're using a bespoke Dockerfile? Probably
not (at the least, we would need to do some refactoring to be able to
discover what path they've overridden with). But maybe a warning if we
discover they're not using the (latest) boilerplate backing image.
- Update convention's `README.md` and `update` output.
- Tests?