golang-osd-operator: Cap concurrency for codecov
Some consumers were regularly having their coverage CI pods OOM-killed.
Experimenting with the prow configuration, we found we needed to set the
memory limits unreasonably high to make it work.

By default, cover tool runs with a number of threads equal to the
detected number of CPUs. In the CI environment at this time, that
appears to be 16. With this commit, we cap it at 4, which brings the
memory usage down to a manageable level without unduly affecting the
runtime of the job.