Fix container-make for openapi-gen
`openapi-gen` is really picky about how `GOPATH` is set with respect to
the path to the repo clone it's working on. And CI is really picky about
where things like `go mod download` are allowed to write. With this
change, we make sure the path inside `container-make`'s container gels
with `GOPATH` so `openapi-gen` works as expected.

Co-Authored-By: Kirk Bater <kbater@redhat.com>