Generate operator bundle generically and inclusively
This is an almost-total rewrite of the common operator bundle generator
script to accommodate variations in boilerplate consumers.

Before: The script looked for specific files by name. This wasn't a big
deal as long as consumers had the right corresponding file they could
simply rename. However, it would blow up if those files weren't present
or didn't contain exactly what was expected; or (importantly) if more
than one file of that kind was needed.

After: The script processes all yaml documents under the deploy/
subdirectory, regardless of filename, including cases where multiple
yaml documents exist in a single file, and processes them based on their
Kind. The following are treated specially:

- CustomResourceDefinitions: These are registered as "owned" in the CSV.
Otherwise they are copied into the bundle directory as usual.
- ClusterRoles: These are folded into the CSV's `clusterPermissions` and
*not* copied into the bundle directory.
- ClusterRoleBindings: These are only used to discover the appropriate
`serviceAccountName` for their corresponding ClusterRoles. They are not
copied into the bundle directory. If any ClusterRoleBindings are "left
over" at the end (not used to find a ServiceAccount), an error is
raised.
- Deployment: We explicitly validate that there is exactly one of these.
We fold the contents into the CSV's `deployments`. The Deployment is
*not* copied to the bundle directory.
- ServiceAccounts: These are *ignored*, as we assume they correspond
exactly to those listed in ClusterRoleBindings or the Deployment's Pod
spec.

TODO:s
- The CSV
[schema](https://docs.openshift.com/container-platform/4.5/rest_api/operatorhub_apis/clusterserviceversion-operators-coreos-com-v1alpha1.html)
appears to support Role[Binding]s in a manner similar to
ClusterRole[Binding]s. Assuming OLM supports that, we should use it. For
now, this commit treats Role[Binding]s as generic resources and simply
copies them into the bundle, as consumers appeared to be doing.
- We should validate ServiceAccounts to make sure they correspond to
[Cluster]RoleBindings or Pod specs in the Deployment, and error if any
are "left over".