Fix CI: Support disabling SaaS file checks
The 06-csv-generate test invokes (make targets that invoke)
csv-generate.sh, which now queries the operator's SaaS file in
app-interface to determine whether it's a hive or cluster operator.
Since the test uses a dummy project, it has no such SaaS file in
app-interface. Even if it did, the CI image doesn't have the right certs
to query it over https.

This commit enables an environment variable to bypass that whole section
of csv-generate.sh, and enables it in the test case.