image-v1.0.1: Ratchet base image to 1.0.0
To make presumbit CI faster, this commit ratchets up the CI image build
to "start" from the previously-released image, `image-v1.0.0`.

The original build.sh is renamed and a fresh build.sh is introduced. The
latter is currently a no-op; subsequent image releases can add to it.

An app-sre-specific Dockerfile combines all the build scripts to
do a full build from scratch in the appsre pipeline.

There is no functional change to the image itself.

OSD-7253