Adds .gitattributes to unacceptable_deltas "ignores" list
With a recent update to boilerplate, a .gitattributes file is created
during the bootstrap process.  This is causing failures in the
"framework/04-update-from-master-and-revert" test, which currently
doesn't expect the .gitattributes file to exist, uncommitted, after the
test bootstrap.

This adds the .gitattributes file to the "ignores" list in the
unacceptable_deltas function of boilerplate/update.

Also runs `./boilerplate/_lib/boilerplate-commit` to "clean" before the
next test run.

Signed-off-by: Christopher Collins <collins.christopher@gmail.com>