image-v2.0.0: Get rid of operator-sdk generate
Remove invocations of and support for `operator-sdk generate`, replacing
these with the corresponding `controller-gen` calls in the `op-generate`
target of openshift/golang-osd-operator's standard.mk.

The operator-sdk-generate.sh helper script is removed. We invoke
controller-gen directly from the `make` target.

We need to preserve and add to the post-CRD-gen `yq` hacks to produce
CRDs compatible with both v3 and v4. These should be able to go away
once 3.11 is dead. (At that time we'll also need to flag controller-gen
to produce CRD v1 instead of v1beta1.)

As written, this will work for consumers whether their APIs are packaged
separately (as in e.g.
https://github.com/openshift/aws-account-operator/pull/580) or not.

With this commit, we produce a fresh backing image that omits the
operator-sdk binaries, but is otherwise the same as image-v1.0.1.

With this commit, we're rebuilding the backing image from scratch, so
the prow and jenkins Dockerfiles are the same. A subsequent commit will
ratchet the former to be based on image-v2.0.0 to speed up builds, in a
spirit similar to #164.

OSD-7352
OSD-7353