Add `make subscriber-report`
Create a `make` target that runs subscriber reports. The intent is to
run this in a postsubmit prow job, the results of which will
(eventually) be posted somewhere/somehow.

Currently only runs `onboarding` and `release` subcommands, as the `pr`
subcommand requires the `gh` CLI to be installed and authenticated.

Part of OSD-5962