Skip update/revert test case if at master
If we're already at master (which happens e.g. during rehearsals on CI
config updates), the 04-update-from-master-and-revert test is a) silly,
and b) going to fail when attempting `boilerplate-commit` because
there's nothing to update and therefore nothing to commit. Add logic to
short out of this test case in this scenario.