golang-osd-operator appsre: Fix docker login
When stealing (ahem, "reappropriating") podman/docker-accommodating
App-SRE pipeline code when hive switched to a rhel8 jenkins node, we
found a latent bug on the docker side: the `REGISTRY_AUTH_FILE` variable
needs to point to an actual file, not just the directory in which the
file is located. Contributing the fix back "upstream".