Enable GOEXPERIMENT=strictfipsruntime
This ensures that the binary will fail to run if running in a non-FIPS
Compliant environment when built with FIPS_ENABLED=true.

GOEXPERIMENT=strictfipsruntime is not supported by Go generally and is
something that Red Hat is supporting in our own fork. Red Hat's fork of
Go is currently in
registry.ci.openshift.org/ocp/builder:rhel-8-golang-1.19-openshift-4.12
which will allow this to work. When building locally, developers should
set FIPS_ENABLED=false to get around this.

Therefore, also rebuilding this image which has the patched Go
containing remediations for CVE-2023-3089

Signed-off-by: Michael Shen <mshen@redhat.com>