Do not create a bundle that replaces itself.
In certain cases it can happen that this tries to create a bundle that
replaces itself, which will lead to installation failures.

As commits that create the same version should not be allowed in
csv-generate.sh by checking for reverts, it should be safe to simply
error out, as this seems to only happen, when an existing bundle gets
updated with the incorrect replace, after it initially had the correct
'replaces' statement.

This commit adds two checks (and the one in the python code should not
really ever be hit):

- Make sure that a bundle that already exists is not regenerated in the
shell script.
- Error out in the python script that actually writes the files if the
previous_version == current_version (if this gets called somewhere else
the initial check might not be in place).