common: image_exists_in_repo check stdout
a newer version of skopeo changed loggers which had an effect on the
stderr now being written to stdout. Update the conditions to check
stdout to allow newer versions to work as expected.

```
❯ podman run docker://quay.io/skopeo/stable:v1.7.0 inspect docker://quay.io/bpratt/image-does-not-exist:latest 2>/dev/null
❯ skopeo --version
skopeo version 1.15.0
❯ skopeo inspect docker://quay.io/bpratt/image-does-not-exist:latest 2>/dev/null
FATA[0000] Error parsing image name "docker://quay.io/bpratt/image-does-not-exist:latest": reading manifest latest in quay.io/bpratt/image-does-not-exist: unauthorized: access to the requested resource is not authorized
```

Signed-off-by: Brady Pratt <bpratt@redhat.com>