config: dl yq from gh releases
go mod cache downloads the entire git repo to build the project, and
this cache stays in the final image. When scanned by `Syft`, it finds
the github workflow files for `yq`. This then generates a garbage
package URL which causes a false positive.

This is likely a bug in Syft, but we can work around it for now by
downloading directly from GH releases.

https://redhat-internal.slack.com/archives/C04PZ7H0VA8/p1722411065681639?thread_ts=1722365748.535409&cid=C04PZ7H0VA8

Signed-off-by: Brady Pratt <bpratt@redhat.com>