KONFLUX-3729: install yq from konflux build
We are running into a `syft` bug/configuration issue that causes a
github action file from yq to be flagged during scanning, blocking the
release. This is due to building from source which clones and caches the
repository in gomod cache. Instead, we can install from a konflux build
of yq which is just a patch version ahead (v4.44.3)

Signed-off-by: Brady Pratt <bpratt@redhat.com>