golang-osd-operator-osde2e: mv e2e tests
on update, move the tests to tests/e2e

if the osde2e directory does not exist, exit
if the test/osde2e directory already exists, exit
otherwise move the files around.

Once everyone is updated, we can drop this code

also update the generation to point to this new location by default in
case there are any left to generate.

Signed-off-by: Brady Pratt <bpratt@redhat.com>