Remove the `GOLANGCI_OPTIONAL_CONFIG` option, which is unused In https://github.com/openshift/boilerplate/pull/92 we added an option to be able to pass a custom golangci-lint configuration on a per repo. I believe the intent is to have it be _additive_, in that it would run a second `golangci-lint` invocation with your custom config. This would only work if you added a linter, as removing one in this file still meant the original invocation would have failed with the boilerplate-based config.
Additionally, the gate that was designed to prevent `golangci-lint` from running a second time, unless `GOLANGCI_OPTIONAL_CONFIG` is set, appears to not be working, so we are currently running linting twice on all our repos.

Lastly, after doing a quick search of GitHub, it appears we only have to repos with the configuration:
- https://github.com/openshift/osd-metrics-exporter/ configures it but the file was previously deleted, so no-ops
- https://github.com/openshift/osd-cluster-ready configures it, but the file just enables `gosec` that is part of our base config

For the above reasons, this PR proposes removing this feature to simplify linting versus fixing bugs with the existing, unused process.