Stop forcing `GOOS=linux` in `go-build` target ~5 years ago this line was modified to enforce linux for building binaries.
This has the side effect of breaking `CGO` compilation on macOS. It can be fixed by
configuring a cross-sompiler that generates objects with Linux header files, however
when building on your _local_ machine (outside docker) there is no need to force this to linux.

When the binary is built inside a container, the `GOOS` will be set to linux.